+++ tests/test_marshmallow.py
@@ -1,9 +1,11 @@
-import pytest
 from datetime import datetime
+
+import pytest
+
 from bson import ObjectId
 import marshmallow
 
-from umongo import Document, EmbeddedDocument, fields, set_gettext, validate
+from umongo import Document, EmbeddedDocument, fields, set_gettext, validate, missing
 from umongo import marshmallow_bonus as ma_bonus_fields
 from umongo.abstract import BaseField, BaseSchema
 from umongo.marshmallow_bonus import (
@@ -133,6 +135,59 @@ class TestMarshmallow(BaseTest):
         assert ma_custom_base_schema.Meta.exclude == ('content',)
         assert ma_custom_base_schema.Meta.dump_only == ('id',)
 
+    def test_as_marshmallow_field_pass_params(self):
+        @self.instance.register
+        class MyDoc(Document):
+            lf = fields.IntField(marshmallow_load_from='lflf')
+            dt = fields.IntField(marshmallow_dump_to='dtdt')
+            at = fields.IntField(marshmallow_attribute='atat')
+            re = fields.IntField(marshmallow_required=True)
+            an = fields.IntField(marshmallow_allow_none=True)
+            lo = fields.IntField(marshmallow_load_only=True)
+            do = fields.IntField(marshmallow_dump_only=True)
+            va = fields.IntField(marshmallow_validate=validate.Range(min=0))
+            em = fields.IntField(marshmallow_error_messages={'invalid': 'Wrong'})
+
+        MyMaDoc = MyDoc.schema.as_marshmallow_schema()
+
+        assert MyMaDoc().fields['lf'].load_from == 'lflf'
+        assert MyMaDoc().fields['dt'].dump_to == 'dtdt'
+        assert MyMaDoc().fields['at'].attribute == 'atat'
+        assert MyMaDoc().fields['re'].required is True
+        assert MyMaDoc().fields['an'].allow_none is True
+        assert MyMaDoc().fields['lo'].load_only is True
+        assert MyMaDoc().fields['do'].dump_only is True
+        _, err = MyMaDoc().load({'va': -1})
+        assert 'va' in err
+        assert MyMaDoc().fields['em'].error_messages['invalid'] == 'Wrong'
+
+    def test_as_marshmallow_field_infer_missing_default(self):
+        @self.instance.register
+        class MyDoc(Document):
+            de = fields.IntField(default=42)
+            mm = fields.IntField(marshmallow_missing=12)
+            md = fields.IntField(marshmallow_default=12)
+            mmd = fields.IntField(default=42, marshmallow_missing=12)
+            mdd = fields.IntField(default=42, marshmallow_default=12)
+
+        MyMaDoc = MyDoc.schema.as_marshmallow_schema()
+
+        data, _ = MyMaDoc().load({})
+        assert data == {
+            'de': 42,
+            'mm': 12,
+            'mmd': 12,
+            'mdd': 42,
+        }
+
+        data, _ = MyMaDoc().dump({})
+        assert data == {
+            'de': 42,
+            'md': 12,
+            'mmd': 42,
+            'mdd': 12,
+        }
+
     def test_as_marshmallow_schema_cache(self):
         ma_schema_cls = self.User.schema.as_marshmallow_schema()
 
@@ -167,7 +222,7 @@ class TestMarshmallow(BaseTest):
         class Vehicle(Document):
             brand = fields.StrField(description='Manufacturer name')
             category = fields.StrField(required=True)
-            nb_wheels = fields.IntField(missing=4)
+            nb_wheels = fields.IntField(default=4)
 
         ma_schema_cls = Vehicle.schema.as_marshmallow_schema()
         schema = ma_schema_cls()
@@ -207,7 +262,7 @@ class TestMarshmallow(BaseTest):
 
         ret = schema.dump({'is_left_handed': True})
         assert not ret.errors
-        assert ret.data == {'name': '1337', 'is_left_handed': True}
+        assert ret.data == {'name': '1337', 'is_left_handed': True, 'cls': 'AdvancedUser'}
 
     def test_to_mongo(self):
         @self.instance.register
@@ -252,8 +307,34 @@ class TestMarshmallow(BaseTest):
         assert ret.data == {'name': 'John'}
 
     def test_missing_accessor(self):
-        # TODO
-        pass
+
+        @self.instance.register
+        class WithDefault(Document):
+            with_umongo_default = fields.StrictDateTimeField(default=datetime(1999, 1, 1))
+            with_marshmallow_missing = fields.StrictDateTimeField(marshmallow_missing='2000-01-01T00:00:00+00:00')
+            with_marshmallow_default = fields.StrictDateTimeField(marshmallow_default='2001-01-01T00:00:00+00:00')
+            with_marshmallow_and_umongo = fields.StrictDateTimeField(
+                default=datetime(1999, 1, 1), marshmallow_missing='2000-01-01T00:00:00+00:00',
+                marshmallow_default='2001-01-01T00:00:00+00:00')
+            with_force_missing = fields.StrictDateTimeField(
+                default=datetime(2001, 1, 1), marshmallow_missing=missing, marshmallow_default=missing)
+            with_nothing = fields.StrField()
+
+        ma_schema = WithDefault.schema.as_marshmallow_schema()()
+        ret = ma_schema.dump({})
+        assert not ret.errors
+        assert ret.data == {
+            'with_umongo_default': '1999-01-01T00:00:00+00:00',
+            'with_marshmallow_default': '2001-01-01T00:00:00+00:00',
+            'with_marshmallow_and_umongo': '2001-01-01T00:00:00+00:00',
+        }
+        ret = ma_schema.load({})
+        assert not ret.errors
+        assert ret.data == {
+            'with_umongo_default': datetime(1999, 1, 1),
+            'with_marshmallow_missing': datetime(2000, 1, 1),
+            'with_marshmallow_and_umongo': datetime(2000, 1, 1),
+        }
 
     def test_nested_field(self):
         @self.instance.register
@@ -330,11 +411,10 @@ class TestMarshmallow(BaseTest):
                 ('id', ma_bonus_fields.ObjectId),
                 ('ref', ma_bonus_fields.ObjectId),
                 ('gen_ref', ma_bonus_fields.GenericReference)
-                ):
+        ):
             ma_field = Doc.schema.fields[name].as_marshmallow_field()
             assert isinstance(ma_field, field_cls)
             assert not isinstance(ma_field, BaseField)
-
 
         oo_data = {
             'id': ObjectId("57c1a71113adf27ab96b2c4f"),