+++ umongo/abstract.py
@@ -110,7 +110,48 @@ class BaseField(ma_fields.Field):
     }
 
     def __init__(self, *args, io_validate=None, unique=False, instance=None, **kwargs):
+        if 'missing' in kwargs:
+            raise RuntimeError("uMongo doesn't use `missing` argument, use `default` "
+                "instead and `marshmallow_missing`/`marshmallow_default` "
+                "to tell `as_marshmallow_field` to use a custom value when "
+                "generating pure Marshmallow field.")
+        if 'default' in kwargs:
+            kwargs['missing'] = kwargs['default']
+
+        # Store attributes prefixed with marshmallow_ to use them when
+        # creating pure marshmallow Schema
+        for attribute in (
+            'load_from', 'dump_to', 'attribute',
+            'validate', 'required', 'allow_none',
+            'load_only', 'dump_only', 'error_messages'
+        ):
+            attribute = 'marshmallow_' + attribute
+            if attribute in kwargs:
+                setattr(self, attribute, kwargs.pop(attribute))
+
+        # Infer from "default" parameter a default value for
+        # marshmallow_default and marshmallow_missing
+        def serialize_default():
+            val = self.default() if callable(self.default) else self.default
+            return self.serialize('foo', {'foo': val})
+
+        self.marshmallow_missing = kwargs.pop('marshmallow_missing', serialize_default)
+        self.marshmallow_default = kwargs.pop('marshmallow_default', serialize_default)
+
         super().__init__(*args, **kwargs)
+
+        # Deserialize default/missing values
+        # This ensures they are validated and get the proper types and constraints
+        for attr in ('default', 'missing'):
+            default = getattr(self, attr)
+            if default is not missing:
+                if callable(default):
+                    def call_default():
+                        return self.deserialize(default())
+                    setattr(self, attr, call_default)
+                else:
+                    setattr(self, attr, self.deserialize(default))
+
         # Overwrite error_messages to handle i18n translation
         self.error_messages = I18nErrorDict(self.error_messages)
         # `io_validate` will be run after `io_validate_resursive`
@@ -126,7 +167,9 @@ class BaseField(ma_fields.Field):
                 'attribute={self.attribute!r}, '
                 'validate={self.validate}, required={self.required}, '
                 'load_only={self.load_only}, dump_only={self.dump_only}, '
-                'missing={self.missing}, allow_none={self.allow_none}, '
+                'marshmallow_missing={self.marshmallow_missing}, '
+                'marshmallow_default={self.marshmallow_default}, '
+                'allow_none={self.allow_none}, '
                 'error_messages={self.error_messages}, '
                 'io_validate={self.io_validate}, '
                 'io_validate_recursive={self.io_validate_recursive}, '
@@ -197,12 +240,26 @@ class BaseField(ma_fields.Field):
         return {self.attribute or key: query}
 
     def _extract_marshmallow_field_params(self, mongo_world):
-        params = {field: getattr(self, field)
-                  for field in ('default', 'load_from', 'validate',
-                                'required', 'allow_none', 'load_only',
-                                'dump_only', 'missing', 'error_messages')}
+        params = {
+            attribute: getattr(self, attribute)
+            for attribute in (
+                'validate', 'required', 'allow_none',
+                'dump_only', 'load_only', 'error_messages'
+            )
+        }
         if mongo_world and self.attribute:
             params['attribute'] = self.attribute
+
+        # Override uMongo attributes with marshmallow_ prefixed attributes
+        for attribute in (
+            'default', 'missing', 'load_from', 'dump_to', 'attribute',
+            'validate', 'required', 'allow_none',
+            'load_only', 'dump_only', 'error_messages'
+        ):
+            ma_attribute = 'marshmallow_' + attribute
+            if hasattr(self, ma_attribute):
+                params[attribute] = getattr(self, ma_attribute)
+
         params.update(self.metadata)
         return params