+++ umongo/document.py
@@ -1,7 +1,6 @@
 from copy import deepcopy
 
 from bson import DBRef
-
 from marshmallow import pre_load, post_load, pre_dump, post_dump, validates_schema  # republishing
 
 from .abstract import BaseDataObject
@@ -315,7 +314,7 @@ class DocumentImplementation(BaseDataObject, Implementation, metaclass=MetaDocum
         """
         Overload this method to get a callback before document insertion.
 
-        .. note:: If you use an async driver, this callback can return a defer/future.
+        .. note:: If you use an async driver, this callback can be asynchronous.
         """
         pass
 
@@ -325,7 +324,7 @@ class DocumentImplementation(BaseDataObject, Implementation, metaclass=MetaDocum
         :return: Additional filters dict that will be used for the query to
             select the document to update.
 
-        .. note:: If you use an async driver, this callback can return a defer/future.
+        .. note:: If you use an async driver, this callback can be asynchronous.
         """
         pass
 
@@ -335,7 +334,7 @@ class DocumentImplementation(BaseDataObject, Implementation, metaclass=MetaDocum
         :return: Additional filters dict that will be used for the query to
             select the document to update.
 
-        .. note:: If you use an async driver, this callback can return a defer/future.
+        .. note:: If you use an async driver, this callback can be asynchronous.
         """
         pass
 
@@ -344,7 +343,7 @@ class DocumentImplementation(BaseDataObject, Implementation, metaclass=MetaDocum
         Overload this method to get a callback after document insertion.
         :param ret: Pymongo response sent by the database.
 
-        .. note:: If you use an async driver, this callback can return a defer/future.
+        .. note:: If you use an async driver, this callback can be asynchronous.
         """
         pass
 
@@ -353,7 +352,7 @@ class DocumentImplementation(BaseDataObject, Implementation, metaclass=MetaDocum
         Overload this method to get a callback after document update.
         :param ret: Pymongo response sent by the database.
 
-        .. note:: If you use an async driver, this callback can return a defer/future.
+        .. note:: If you use an async driver, this callback can be asynchronous.
         """
         pass
 
@@ -362,6 +361,6 @@ class DocumentImplementation(BaseDataObject, Implementation, metaclass=MetaDocum
         Overload this method to get a callback after document deletion.
         :param ret: Pymongo response sent by the database.
 
-        .. note:: If you use an async driver, this callback can return a defer/future.
+        .. note:: If you use an async driver, this callback can be asynchronous.
         """
         pass